# swagger.yaml
openapi: 3.0.3
info:
  title: Binance Proxy API
  description: |
    Proxy para a API da Binance que permite acesso CORS e facilita integração frontend.
    Este proxy repassa todas as requisições para a API oficial da Binance (https://api.binance.com/api/v3).
    
    **Endpoints do Proxy:**
    - `/health` - Verifica se o proxy está funcionando
    - `/test` - Testa a conexão com a Binance
    
    **Endpoints da Binance (proxied):**
    Todos os endpoints da API v3 da Binance estão disponíveis através deste proxy.
  version: 1.0.0
  contact:
    name: Binance Proxy Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Servidor local de desenvolvimento
  - url: https://api.binance.com/api/v3
    description: API oficial da Binance (referência)
...
paths:
...
  /test:
    get:
      tags:
        - Proxy
      summary: Test Connection
      description: Testa a conexão com a API da Binance
      operationId: testConnection
      responses:
        '200':
          description: Conexão estabelecida com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  binance_url:
                    type: string
                    example: https://api.binance.com/api/v3
                  http_status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Conexão com Binance estabelecida com sucesso
        '503':
          description: Erro ao conectar com Binance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: Código de erro
          example: -1000
        msg:
          type: string
          description: Mensagem de erro
          example: "Erro ao conectar com Binance"
        message:
          type: string
          description: Mensagem de erro detalhada
          example: "Erro ao conectar com Binance: connection timeout"
...